pyproject.toml
[project]
name = "line-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "line-bot-sdk>=3.21.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
